# sidecar/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CodebookEntryOut": {
        "description": "Ein Eintrag im Codebook (Kategorie oder Subcode).",
        "properties": {
          "anchors": {
            "items": {
              "type": "string"
            },
            "title": "Anchors",
            "type": "array"
          },
          "code_id": {
            "title": "Code Id",
            "type": "string"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hex-Farbe, z.B. '#ffcc00'.",
            "title": "Color"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "code_id",
          "name"
        ],
        "title": "CodebookEntryOut",
        "type": "object"
      },
      "ExportRequest": {
        "additionalProperties": false,
        "description": "Export-Request fuer QDPX oder XLSX.",
        "properties": {
          "include_annotations": {
            "default": true,
            "title": "Include Annotations",
            "type": "boolean"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Leer = letzter Run.",
            "title": "Run Id"
          }
        },
        "required": [
          "project_id"
        ],
        "title": "ExportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Lokaler FastAPI-Sidecar fuer die Qualdatan-Desktop-App. Contract-first: diese OpenAPI-Spec ist in Phase F gefroren; Welle 1 implementiert die Endpoint-Koerper, Welle 2 generiert den Frontend-Client.",
    "title": "Qualdatan Sidecar",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/codebook/{project_id}": {
      "get": {
        "description": "Liefert das zusammengefuehrte Codebook eines Projekts.",
        "operationId": "get_codebook_codebook__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CodebookEntryOut"
                  },
                  "title": "Response Get Codebook Codebook  Project Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Codebook",
        "tags": [
          "codebook"
        ]
      }
    },
    "/codebook/{project_id}/codes/{code_id}": {
      "delete": {
        "description": "Entfernt einen Codebook-Eintrag (nur Override bzw. Projekt-lokal).",
        "operationId": "delete_codebook_entry_codebook__project_id__codes__code_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "code_id",
            "required": true,
            "schema": {
              "title": "Code Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Codebook Entry",
        "tags": [
          "codebook"
        ]
      },
      "patch": {
        "description": "Legt ein Projekt-Override fuer einen Codebook-Eintrag an.",
        "operationId": "patch_codebook_entry_codebook__project_id__codes__code_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "code_id",
            "required": true,
            "schema": {
              "title": "Code Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeOverridePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodebookEntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Codebook Entry",
        "tags": [
          "codebook"
        ]
      }
    },
    "/export/qdpx": {
      "post": {
        "description": "Exportiert einen Run als REFI-QDA (QDPX).",
        "operationId": "export_qdpx_export_qdpx_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Qdpx",
        "tags": [
          "export"
        ]
      }
    },
    "/export/xlsx": {
      "post": {
        "description": "Exportiert kodierte Segmente eines Runs als XLSX.",
        "operationId": "export_xlsx_export_xlsx_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Xlsx",
        "tags": [
          "export"
        ]
      }
    },
...
    "/plugins": {
      "get": {
        "description": "Listet lokal installierte Plugin-Bundles.",
        "operationId": "list_plugins_plugins_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundleSummary"
                  },
                  "title": "Response List Plugins Plugins Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/plugins/available": {
      "get": {
        "description": "Listet im Plugin-Server verfuegbare (noch nicht installierte) Bundles.",
        "operationId": "list_available_plugins_plugins_available_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundleAvailable"
                  },
                  "title": "Response List Available Plugins Plugins Available Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Available Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/plugins/install": {
      "post": {
        "description": "Installiert ein Bundle aus dem Plugin-Server.",
        "operationId": "install_plugin_plugins_install_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/plugins/{bundle_id}/uninstall": {
      "post": {
        "description": "Deinstalliert ein lokal installiertes Bundle.",
        "operationId": "uninstall_plugin_plugins__bundle_id__uninstall_post",
        "parameters": [
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uninstall Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/projects": {
      "get": {
        "description": "Listet alle Projekte.",
        "operationId": "list_projects_projects_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response List Projects Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "Legt ein neues Projekt an.",
        "operationId": "create_project_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}": {
      "delete": {
        "description": "Loescht ein Projekt (inkl. aller Runs, Materialien, Overrides).",
        "operationId": "delete_project_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "description": "Liefert ein Projekt per ID.",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "description": "Teil-Update eines Projekts.",
        "operationId": "patch_project_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/plugins/{bundle_id}/disable": {
      "post": {
        "description": "Deaktiviert ein Bundle fuer ein Projekt.",
        "operationId": "disable_plugin_for_project_projects__project_id__plugins__bundle_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Plugin For Project",
        "tags": [
          "plugins"
        ]
      }
    },
    "/projects/{project_id}/plugins/{bundle_id}/enable": {
      "post": {
        "description": "Aktiviert ein Bundle fuer ein Projekt.",
        "operationId": "enable_plugin_for_project_projects__project_id__plugins__bundle_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Sidecar-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sidecar-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Plugin For Project",
        "tags": [
          "plugins"
        ]
      }
    },
...
  }
}